Move placeholder component to ui folder
