+++ web/components/mint.tsx
@@ -23,7 +23,7 @@ import { tokenList } from '@components/token-swap-list';
 import { VaultProps } from './vault';
 import { EmptyState } from './ui/empty-state';
 import { ArchiveIcon } from '@heroicons/react/outline';
-import { Placeholder } from './placeholder';
+import { Placeholder } from './ui/placeholder';
 import { InformationCircleIcon } from '@heroicons/react/solid';
 import { Tooltip } from '@blockstack/ui';
