+++ web/components/swap.tsx
@@ -22,7 +22,7 @@ import { TokenSwapList, tokenList } from '@components/token-swap-list';
 import { SwapSettings } from '@components/swap-settings';
 import { getBalance } from '@components/app';
 import { classNames } from '@common/class-names';
-import { Placeholder } from './placeholder';
+import { Placeholder } from './ui/placeholder';
 import { SwapLoadingPlaceholder } from './swap-loading-placeholder';
 
 export const Swap: React.FC = () => {
